[TM-3193] fix commentaries on PR
